- tested all function of hardware

+++ 1. MicroMouse Code/2. MicroMouse - STM32F407VGT6/PCB_AIO_V2_CODE/Core/Inc/AML_PID.h
@@ -6,9 +6,9 @@
 
 typedef struct
 {
-    uint32_t *MyInput;
-    uint32_t *MyOutput;
-    uint32_t *MySetpoint;
+    double *MyInput;
+    double *MyOutput;
+    double *MySetpoint;
 
     double Kp;
     double Ki;
@@ -33,7 +33,7 @@ typedef struct
     double out;
 } AML_PID_Struct;
 
-void AML_PID_Init(AML_PID_Struct *pid, uint32_t *input, uint32_t *output, uint32_t *setpoint, double kp, double ki, double kd, double tau, double limMin, double limMax, double linMinInt, double linMaxInt, uint32_t sampleTime);
+void AML_PID_Init(AML_PID_Struct *pid, double *input, double *output, double *setpoint, double kp, double ki, double kd, double tau, double limMin, double limMax, double linMinInt, double linMaxInt, uint32_t sampleTime);
 double AML_PID_Compute(AML_PID_Struct *pid);
 
 #endif // AML_PID_H
